Refactor two 1000+ line files into focused submodules
Split the two largest source files into directory modules, each grouped
by concern. Pure mechanical decomposition — no behaviour change.

mogen-studio/src/app/ui_wizard.rs (2406 lines) -> ui_wizard/:
  - mod.rs      WizardSession/WizardBusy + batch-claim helpers
  - session.rs  window lifecycle, poll, draw dispatch, action routing
  - workers.rs  background-thread launches for each pipeline stage
  - draw.rs     egui stage renderers + WizardAction enum
  - message.rs  folding worker results into the session
  - tests.rs    batch claim/reservation unit tests

mogen-dsl/.../emit/circulation.rs (1099 lines) -> circulation/:
  - mod.rs       emit_circulation entry, column fillers, shared
                 constants/FlightDir/scene-graph helpers
  - stairs.rs    switchback flights, treads, mid-landings
  - handrails.rs sloped + flat railings
  - elevator.rs  shaft enclosure walls + per-storey west pieces

Largest resulting file is 840 lines. Full workspace builds clean;
mogen-studio ui_wizard tests (8) and mogen-dsl building tests (89) pass.
